requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
gunicorn==20.0.4
python-dotenv==1.0.0
psycopg2-binary==2.9.7
pyTelegramBotAPI==4.10.0
requests==2.32.3